src/pyproject.toml
[tool.poetry]
name = "monitoring-test-task"
version = "0.1.0"
description = ""
authors = ["Rashid Mingazov <my@rdmingazov.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
pydantic = "^1.9.2"
aiohttp = "^3.8.1"
uvicorn = "^0.18.2"
loguru = "^0.6.0"
sqlmodel = "^0.0.6"
asyncpg = "^0.26.0"
greenlet = "^1.1.2"
alembic = "^1.8.1"
pyhumps = "^3.8.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^5.0.4"
isort = "^5.10.1"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
